refactor(progress-bar): обновлять стиль через состояние
